backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
supabase==2.0.0
python-multipart==0.0.6
PyJWT==2.8.0
websockets>=12.0,<13.0
google-generativeai==0.3.2 
